chore: Partial
