+++ crates/server/src/lib.rs
@@ -5,6 +5,7 @@ pub mod api;
 pub mod audit;
 pub mod build_info;
 pub mod builder;
+pub mod config;
 pub mod dashboard;
 pub mod middleware;
